fix: Show meaningful line numbers for render errors
Our developers found the error messages from renders vague and unhelpful
because they did not identify the source of the error. Fix this by
printing the stacktrace. The code was not preserving stacktraces when
re-throwing exceptions, which meant that if we just printed the traces
as they were, the line from the erb would not be mentioned.

In addition, the line numbers reported are not useful without looking at
the merged template, so number that template on output.